refactor: 파일 내에서 각 기능 별로 분리
파일 내에서 세분화적으로 각 기능을 분리하였습니다.
