+++ scripts/all/allJs.js
@@ -1,5 +1,5 @@
-import "/Users/an-yelim/codesquad/fe--amazon/scripts/nav/navDimmed.js";
-import "/Users/an-yelim/codesquad/fe--amazon/scripts/sidebar/sidebarDimmed.js";
+import "../nav/dimmedNav.js";
+import "../sidebar/dimmedSidebar.js";
 import "/Users/an-yelim/codesquad/fe--amazon/scripts/sidebar/showSidebar.js";
 import "/Users/an-yelim/codesquad/fe--amazon/scripts/sidebar/sidebarInputData.js";
 import "/Users/an-yelim/codesquad/fe--amazon/scripts/all/allQuery.js";
