+++ scripts/all/allQuery.js
@@ -1,12 +1,5 @@
-const ALL_SELECTORS = {
-  dimmed: document.querySelector(".dim-layer"),
-  sidebarAll: document.querySelector(".sub__all-menu"),
-  sidebarClose: document.querySelector(".sidebar__btn_close"),
-  loginMenu: document.querySelector(".nav__login_popup-box"),
-  shippingMenu: document.querySelector(".nav__shipping"),
-  sidebar: document.querySelector(".sidebar"),
-  menuAll: document.querySelector(".sidebar__menu-all"),
-  menuAllDetail: document.querySelector(".sidebar__all-menu_extend"),
-};
+const $ = (selector, parent = document) => parent.querySelector(selector);
+const $$ = (selector, parent = document) => parent.querySelectorAll(selector);
 
-export { ALL_SELECTORS };
+export { $, $$ };
+
