+++ scripts/sidebar/dimmedSidebar.js
