+++ styleSheet/nav/navAnimation.css
@@ -6,7 +6,7 @@
     to{
         opacity: 1;
         visibility: visible;
-         }
+    }
 }
 
 .nav__login_popup{
